test: move ledger helpers file to top-level helpers
